feat(profile): unify the RIA advisor profile into the main /profile section
The RIA subagent had a bespoke /ria/profile screen while every other subagent
(e.g. location) shows the unified /profile. Bring RIA into the same section: the
full advisor profile now lives under the /profile "Regulatory profile" panel with
identical UI/UX and all logic (view, edit services, re-initiate, delete, and the
license-data refresh), and the RIA subagent's Profile tab opens the unified page.

- components/ria/profile/ria-profile-section.tsx: extract the former /ria/profile
  body into a shared RiaProfileSection (host passes RIA onboarding status + a
  refresh callback; the section owns all mutations). Folds in the license-refresh
  action so nothing is lost.
- app/profile/profile-workspace-page.tsx: render RiaProfileSection as the
  "regulatory" stack panel; the "Regulatory profile" row now opens that full panel;
  retire the old lightweight inline "Update license data" modal (+ its state,
  handlers, and voice/agent metadata refs).
- lib/navigation/profile-routes.ts: add the deep-linkable "regulatory" panel.
- lib/navigation/app-bottom-nav.ts: Profile tab → unified /profile for ALL scopes
  (RIA no longer routes to the bespoke page).
- Redirect /ria/profile + /ria/settings → /profile?panel=regulatory; point RIA
  onboarding completion and the marketplace "Manage your profile" link there too.
- Tests: retarget the RIA profile tests to RiaProfileSection; update the bottom-nav
  and onboarding-route expectations.

Investor-only accounts are unaffected (the Regulatory row stays RIA-gated).
