+++ hushh-webapp/__tests__/app/ria/profile-page.test.tsx
@@ -10,7 +10,11 @@ const mocks = vi.hoisted(() => ({
   refresh: vi.fn(),
   switchPersona: vi.fn(),
   toast: { success: vi.fn(), error: vi.fn(), info: vi.fn() },
-  riaService: { deleteProfile: vi.fn(), updateProfile: vi.fn() },
+  riaService: {
+    deleteProfile: vi.fn(),
+    updateProfile: vi.fn(),
+    refreshLicenseProfile: vi.fn(),
+  },
   openKaiCommandBar: vi.fn(),
   draftClear: vi.fn(),
   onPersonaStateChanged: vi.fn(),
@@ -23,6 +27,7 @@ vi.mock("next/navigation", () => ({
 
 vi.mock("lucide-react", () => ({
   ArrowRight: () => <span />,
+  ClipboardCheck: () => <span />,
   Loader2: () => <span />,
   RotateCcw: () => <span />,
   Trash2: () => <span />,
@@ -67,6 +72,45 @@ vi.mock("@/components/app-ui/settings-ui", () => ({
   }) => (
     <button data-testid={testId} onClick={onClick} disabled={disabled}>
       {title}
+    </button>
+  ),
+}));
+
+vi.mock("@/components/ui/dialog", () => ({
+  Dialog: ({
+    open,
+    children,
+  }: {
+    open: boolean;
+    children: React.ReactNode;
+  }) => (open ? <div data-testid="license-dialog">{children}</div> : null),
+  DialogContent: ({ children }: { children: React.ReactNode }) => (
+    <div>{children}</div>
+  ),
+  DialogTitle: ({ children }: { children: React.ReactNode }) => (
+    <div>{children}</div>
+  ),
+  DialogDescription: ({ children }: { children: React.ReactNode }) => (
+    <div>{children}</div>
+  ),
+}));
+
+vi.mock("@/components/ui/input", () => ({
+  Input: (props: Record<string, unknown>) => <input {...props} />,
+}));
+
+vi.mock("@/components/ui/button", () => ({
+  Button: ({
+    children,
+    onClick,
+    disabled,
+  }: {
+    children: React.ReactNode;
+    onClick?: () => void;
+    disabled?: boolean;
+  }) => (
+    <button onClick={onClick} disabled={disabled}>
+      {children}
     </button>
   ),
 }));
@@ -148,7 +192,7 @@ vi.mock("@/lib/navigation/kai-command-bar-events", () => ({
   openKaiCommandBar: mocks.openKaiCommandBar,
 }));
 
-import RiaProfilePage from "@/app/ria/profile/page";
+import { RiaProfileSection } from "@/components/ria/profile/ria-profile-section";
 
 const VERIFIED_STATUS = {
   exists: true,
@@ -160,7 +204,20 @@ const VERIFIED_STATUS = {
   fee_structure: ["Hourly"],
 };
 
-describe("RiaProfilePage manage actions", () => {
+// The RIA advisor profile now renders as the shared RiaProfileSection inside the
+// unified /profile "Regulatory profile" panel. The host owns the status fetch and
+// passes it in; the section owns view/edit/re-initiate/delete/license logic.
+function renderSection(status: unknown = VERIFIED_STATUS) {
+  return render(
+    <RiaProfileSection
+      status={status as never}
+      loading={false}
+      onRefresh={vi.fn()}
+    />,
+  );
+}
+
+describe("RiaProfileSection manage actions", () => {
   beforeEach(() => {
     vi.clearAllMocks();
     mocks.useAuth.mockReturnValue({
@@ -183,7 +240,7 @@ describe("RiaProfilePage manage actions", () => {
   });
 
   it("renders the Manage group with re-initiate and delete rows", async () => {
-    render(<RiaProfilePage />);
+    renderSection();
     await waitFor(() => {
       expect(screen.getByTestId("ria-profile-reinitiate")).toBeTruthy();
       expect(screen.getByTestId("ria-profile-delete")).toBeTruthy();
@@ -191,7 +248,7 @@ describe("RiaProfilePage manage actions", () => {
   });
 
   it("re-initiate routes to the onboarding wizard with ?reinitiate=1", async () => {
-    render(<RiaProfilePage />);
+    renderSection();
     await waitFor(() =>
       expect(screen.getByTestId("ria-profile-reinitiate")).toBeTruthy(),
     );
@@ -204,7 +261,7 @@ describe("RiaProfilePage manage actions", () => {
   });
 
   it("delete confirms, calls deleteProfile, switches persona, and routes to One", async () => {
-    render(<RiaProfilePage />);
+    renderSection();
     await waitFor(() =>
       expect(screen.getByTestId("ria-profile-delete")).toBeTruthy(),
     );
@@ -245,7 +302,7 @@ describe("RiaProfilePage manage actions", () => {
       switchPersona: mocks.switchPersona,
     });
 
-    render(<RiaProfilePage />);
+    renderSection({ exists: false });
 
     await waitFor(() =>
       expect(mocks.routerReplace).toHaveBeenCalledWith("/ria/onboarding"),
