+++ hushh-webapp/app/ria/settings/page.tsx
@@ -1,6 +1,6 @@
 import { ClientRedirect } from "@/components/navigation/client-redirect";
-import { ROUTES } from "@/lib/navigation/routes";
+import { buildProfileRoute } from "@/lib/navigation/profile-routes";
 
 export default function RiaSettingsCompatibilityPage() {
-  return <ClientRedirect to={ROUTES.RIA_PROFILE} />;
+  return <ClientRedirect to={buildProfileRoute({ panel: "regulatory" })} />;
 }
